requirements.txt
fastapi
uvicorn
pymongo
dnspython
python-jose
passlib[bcrypt]
python-dotenv
